pyproject.toml
[project]
name = "lgs-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.21.0",
]
